// packages/api/src/api/types/generated.ts
export type ConfigMap = Record<string, string>;
// ...
export interface StorageVolume extends StorageVolumePut {
  /**
   * Volume name
   * @example foo
   */
  name: string;
  /**
   * Volume type
   * @example custom
   */
  type: string;
  /**
   * List of URLs of objects using this storage volume
   * @example ["/1.0/instances/blah"]
   */
  used_by: string[];
  /**
   * What cluster member this record was found on
   * @example server01
   *
   * API extension: clustering
   */
  location: string;
  /**
   * Volume content type (filesystem or block)
   * @example filesystem
   *
   * API extension: custom_block_volumes
   */
  content_type: string;
  /**
   * Project containing the volume.
   * @example default
   *
   * API extension: storage_volumes_all_projects
   */
  project: string;
  /**
   * Volume creation timestamp
   * @example 2021-03-23T20:00:00-04:00
   * API extension: storage_volumes_created_at
   */
  created_at: string;
}
// ...
export interface StorageVolumePut {
  /**
   * Storage volume configuration map (refer to doc/storage.md)
   * @example {"zfs.remove_snapshots": "true", "size": "50GiB"}
   */
  config: ConfigMap;
  /**
   * Description of the storage volume
   * @example My custom volume
   *
   * API extension: entity_description
   */
  description: string;
  /**
   * Name of a snapshot to restore
   * @example snap0
   *
   * API extension: storage_api_volume_snapshots
   */
  restore?: string;
}

/**
 * StorageVolumeRebuildPost represents the fields available for a storage volume rebuild request.
 *
 *
 * API extension: storage_volumes_rebuild.
 */
export type StorageVolumeRebuildPost = unknown;

/**
 * StorageVolumeSnapshot represents a storage volume snapshot
 *
 *
 * API extension: storage_api_volume_snapshots.
 */
export interface StorageVolumeSnapshot extends StorageVolumeSnapshotPut {
  /**
   * Snapshot name
   * @example snap0
   */
  name: string;
  /**
   * Storage volume configuration map (refer to doc/storage.md)
   * @example {"zfs.remove_snapshots": "true", "size": "50GiB"}
   */
  config: ConfigMap;
  /**
   * The content type (filesystem or block)
   * @example filesystem
   *
   * API extension: custom_block_volumes
   */
  content_type: string;
  /**
   * Volume snapshot creation timestamp
   * @example 2021-03-23T20:00:00-04:00
   * API extension: storage_volumes_created_at
   */
  created_at: string;
}
// ...
export interface StorageVolumeSnapshotPut {
  /**
   * Description of the storage volume
   * @example My custom volume
   */
  description: string;
  /**
   * When the snapshot expires (gets auto-deleted)
   * @example 2021-03-23T17:38:37.753398689-04:00
   *
   * API extension: custom_volume_snapshot_expiry
   */
  expires_at: string | null;
}
// ...
export interface StorageVolumeState {
  /**
   * Volume usage
   */
  usage: StorageVolumeStateUsage | null;
}

/**
 * StorageVolumeStateUsage represents the disk usage of a volume
 *
 *
 * API extension: storage_volume_state.
 */
export interface StorageVolumeStateUsage {
  /**
   * Used space in bytes
   * @example 1693552640
   */
  used?: number;
  /**
   * Storage volume size in bytes
   * @example 5189222192
   *
   * API extension: storage_volume_state_total
   */
  total: number;
}